chore: update stories to modern flavor
Reorganize stories locations.
Fix type exports causing trouble to storybook.
